+++ R/tm_t_abnormality_by_worst_grade.R
@@ -150,7 +150,10 @@ template_abnormality_by_worst_grade <- function(parentname, # nolint: object_len
   y$layout_prep <- bracket_expr(prep_list)
 
   parsed_basic_table_args <- teal.widgets::parse_basic_table_args(
-    teal.widgets::resolve_basic_table_args(user_table = basic_table_args)
+    teal.widgets::resolve_basic_table_args(
+      user_table = basic_table_args,
+      module_table = teal.widgets::basic_table_args(show_colcounts = TRUE)
+    )
   )
 
   # layout start
@@ -163,8 +166,7 @@ template_abnormality_by_worst_grade <- function(parentname, # nolint: object_len
           rtables::split_cols_by(
             var = arm_var,
             split_fun = add_overall_level(label = total_label, first = FALSE)
-          ) %>%
-          rtables::add_colcounts(),
+          ),
         env = list(
           arm_var = arm_var,
           total_label = total_label,
@@ -174,8 +176,7 @@ template_abnormality_by_worst_grade <- function(parentname, # nolint: object_len
     } else {
       substitute(
         expr = expr_basic_table_args %>%
-          rtables::split_cols_by(var = arm_var) %>%
-          rtables::add_colcounts(),
+          rtables::split_cols_by(var = arm_var),
         env = list(arm_var = arm_var, expr_basic_table_args = parsed_basic_table_args)
       )
     }